Home

ConstituencyTrees.jl

ConstituencyTrees.jl is a Julia package for working with constituency trees of natural language sentences (also called parse trees, syntax trees).

Trees

ConstituencyTrees.ConstituencyTree — Type.

ConstituencyTree{T}

Constituency parse tree of a natural language sentence.

source

ConstituencyTrees.@tree_str — Macro.

@tree_str(str)

String macro for reading a constituency parse tree from bracketed format.

tree"(S (NP (DT the) (N cat)) (VP (V ate)))"

source

ConstituencyTrees.Brackets.read_tree — Function.

read_tree(str)

source

ConstituencyTrees.pprint — Function.

pprint([io::IO,] tree; depth=0, indent=2, multiline=true)

Print a constituency parse tree in bracketed format.

Arguments

io:

tree: the tree to search

Keywords

depth=0:

indent=2:

multiline=true:

Returns

String:

source

ConstituencyTrees.POS — Function.

POS(tree)

Iterator for part-of-speech tagged words.

julia> POS(tree"(S (NP (DT the) (N cat)) (VP (V ate)))") |> collect
3-element Array{Any,1}:
 ("DT", "the")
 ("N", "cat")
 ("V", "ate")

Arguments

tree: the tree to search

Returns

POSIterator- a lazy iterator over (POS, token) pairs

source

ConstituencyTrees.Words — Function.

Words(tree)

Iterator for words in a sentence.

julia> Words(tree"(S (NP (DT the) (N cat)) (VP (V ate)))") |> collect
3-element Array{Any,1}:
 "the"
 "cat"
 "ate"

Arguments

tree: the tree to search

Returns

WordsIterator: an iterator over tokens

source

ConstituencyTrees.productions — Function.

productions(tree; search=PreOrderDFS, nonterminal=identity, terminal=identity)

Return a vector of (lhs, rhs) productions from a constituency parse tree.

Arguments

tree: the tree to search

Keywords

nonterminal: a function to call on nonterminal symbols.

terminal: a function to call on nonterminal symbols.

Returns

Vector:

source

Transformations

ConstituencyTrees.chomsky_normal_form — Function.

chomsky_normal_form(tree, factor=RightFactored(), labelf)

Convert a tree into chomsky normal form.

Arguments

tree:

factor::Factorization=RightFactored:

labelf:

Returns

a tree of the same type as its argument

source

ConstituencyTrees.collapse_unary — Function.

collapse_unary(tree, labelf=unary_label; collapse_pos=false, collapse_root=false)

Arguments

tree:

labelf:

Keywords

collapse_pos=false:

collapse_root=false:

Returns

a tree of the same type as its argument

source